+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.2.0 (2013-08-09)
+
+Features:
+
+ - Added Read-EntireFile function `alias: .<`
+ - Added Replace-InFile function `alias: .//`
+ - Added Replace-InFiles function `alias: =//`
+
 ## v0.1.0 (2013-08-06)
 
 Initial public release